requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==19.3.0
CacheControl==0.12.11
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
discord==1.7.3
discord.py==1.7.3
firebase-admin==5.2.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.46.0
grpcio-status==1.46.0
httplib2==0.18.1
idna==2.9
idna-ssl==1.1.0
msgpack==1.0.0
multidict==4.7.5
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
